+++ src/dotcom/requests.ts
@@ -1,19 +1,29 @@
-import { BannerPayload, EpicPayload, HeaderPayload } from '../shared/types';
+import {
+    BannerPayload,
+    BannerProps,
+    EpicPayload,
+    EpicProps,
+    HeaderPayload,
+    HeaderProps,
+    GutterPayload,
+    GutterProps,
+} from '../shared/types';
+
 import { TestTracking } from '../shared/types/abTests/shared';
 
-export interface ModuleData {
+export interface ModuleData<PROPS> {
     name: string;
-    props: Record<string, unknown>; // the client doesn't need to know about the module props
+    props: PROPS;
 }
 
-export interface ModuleDataResponse {
+export interface ModuleDataResponse<PROPS> {
     data?: {
-        module: ModuleData;
+        module: ModuleData<PROPS>;
         meta: TestTracking;
     };
 }
 
-type ModuleType = 'epic' | 'liveblog-epic' | 'banner' | 'header';
+type ModuleType = 'epic' | 'liveblog-epic' | 'banner' | 'header' | 'gutter-liveblog';
 
 const getForcedVariant = (type: ModuleType): string | null => {
     if (URLSearchParams) {
@@ -27,13 +37,13 @@ const getForcedVariant = (type: ModuleType): string | null => {
     return null;
 };
 
-type Payload = EpicPayload | BannerPayload | HeaderPayload;
+type Payload = EpicPayload | BannerPayload | HeaderPayload | GutterPayload;
 
-const getModuleData = (
+const getModuleData = <PROPS>(
     type: ModuleType,
     baseUrl: string,
     payload: Payload,
-): Promise<ModuleDataResponse> => {
+): Promise<ModuleDataResponse<PROPS>> => {
     const forcedVariant = getForcedVariant(type);
     const queryString = forcedVariant ? `?force=${forcedVariant}` : '';
     const url = `${baseUrl}/${type}${queryString}`;
@@ -55,16 +65,27 @@ const getModuleData = (
         .then((response) => response.json());
 };
 
-export const getEpic = (baseUrl: string, payload: EpicPayload): Promise<ModuleDataResponse> =>
-    getModuleData('epic', baseUrl, payload);
+export const getEpic = (
+    baseUrl: string,
+    payload: EpicPayload,
+): Promise<ModuleDataResponse<EpicProps>> => getModuleData('epic', baseUrl, payload);
 
 export const getLiveblogEpic = (
     baseUrl: string,
     payload: EpicPayload,
-): Promise<ModuleDataResponse> => getModuleData('liveblog-epic', baseUrl, payload);
+): Promise<ModuleDataResponse<EpicProps>> => getModuleData('liveblog-epic', baseUrl, payload);
 
-export const getBanner = (baseUrl: string, payload: BannerPayload): Promise<ModuleDataResponse> =>
-    getModuleData('banner', baseUrl, payload);
+export const getBanner = (
+    baseUrl: string,
+    payload: BannerPayload,
+): Promise<ModuleDataResponse<BannerProps>> => getModuleData('banner', baseUrl, payload);
 
-export const getHeader = (baseUrl: string, payload: HeaderPayload): Promise<ModuleDataResponse> =>
-    getModuleData('header', baseUrl, payload);
+export const getGutterLiveblog = (
+    baseUrl: string,
+    payload: GutterPayload,
+): Promise<ModuleDataResponse<GutterProps>> => getModuleData('gutter-liveblog', baseUrl, payload);
+
+export const getHeader = (
+    baseUrl: string,
+    payload: HeaderPayload,
+): Promise<ModuleDataResponse<HeaderProps>> => getModuleData('header', baseUrl, payload);