+++ src/shared/types/props/index.ts
@@ -3,3 +3,4 @@ export * from './banner';
 export * from './header';
 export * from './shared';
 export * from './design';
+export * from './gutter';